add fritz file
